requirements.txt
pyautogui>=0.9.54
python-dotenv>=1.0.1
pyinstaller